# api/swagger/swagger.yaml
definitions:
...
  models.Account:
    properties:
      account_type:
        $ref: '#/definitions/models.AccountType'
      create_at:
        type: string
      delete_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      update_at:
        type: string
      user_id:
        type: integer
      username:
        minLength: 3
        type: string
    required:
    - username
    type: object
  models.AccountType:
    enum:
    - admin
    type: string
    x-enum-varnames:
    - Admin
  models.Category:
    properties:
      create_at:
        type: string
      delete_at:
        type: string
      id:
        type: integer
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      update_at:
        type: string
    type: object
  models.Product:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      create_at:
        type: string
      delete_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      update_at:
        type: string
    type: object
  models.Table:
    properties:
      account:
        $ref: '#/definitions/models.Account'
      account_id:
        type: integer
      create_at:
        type: string
      delete_at:
        type: string
      id:
        type: integer
      name:
        type: string
      pos_x:
        type: integer
      pos_y:
        type: integer
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
      update_at:
        type: string
    type: object
  models.Ticket:
    properties:
      account:
        $ref: '#/definitions/models.Account'
      account_id:
        type: integer
      create_at:
        type: string
      delete_at:
        type: string
      id:
        type: integer
      ticket_products:
        items:
          $ref: '#/definitions/models.TicketProduct'
        type: array
      ticket_status:
        $ref: '#/definitions/models.TicketStatus'
      total:
        type: number
      update_at:
        type: string
    type: object
  models.TicketProduct:
    properties:
      create_at:
        type: string
      delete_at:
        type: string
      id:
        type: integer
      product:
        $ref: '#/definitions/models.Product'
      quantity:
        type: integer
      ticket_id:
        type: integer
      update_at:
        type: string
    type: object
  models.TicketStatus:
    enum:
    - open
    type: string
    x-enum-varnames:
    - TicketOpen
  models.User:
    properties:
      account:
        $ref: '#/definitions/models.Account'
      create_at:
        type: string
      delete_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      update_at:
        type: string
    type: object
  models_errors.ErrorResponse:
    properties:
      description:
        type: string
      details:
        type: string
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Point Of Sale - Total Tools
  title: Point Of Sale API
  version: "1.0"
paths:
...
  /api/v1/payments/tickets/{ticketId}:
    post:
      parameters:
      - description: Ticket ID
        in: path
        name: ticketId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
      security:
      - ApiKeyAuth: []
      tags:
      - Payments
...
  /api/v1/tables/{id}/ticket:
    post:
      parameters:
      - description: Table ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Table'
            type: array
      security:
      - ApiKeyAuth: []
      tags:
      - Table
  /api/v1/tickets:
    get:
      parameters:
      - description: Only your tickets
        in: query
        name: mine
        type: boolean
      - description: Only open tickets
        in: query
        name: onlyOpen
        type: boolean
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ticket'
            type: array
      security:
      - ApiKeyAuth: []
      tags:
      - Ticket
    post:
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
      security:
      - ApiKeyAuth: []
      tags:
      - Ticket
  /api/v1/tickets/{id}:
    delete:
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
      security:
      - ApiKeyAuth: []
      tags:
      - Ticket
    get:
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
      security:
      - ApiKeyAuth: []
      tags:
      - Ticket
  /api/v1/tickets/{ticketId}/products/{productId}:
    delete:
      parameters:
      - description: Ticket ID
        in: path
        name: ticketId
        required: true
        type: integer
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
      security:
      - ApiKeyAuth: []
      tags:
      - Ticket
    post:
      parameters:
      - description: Ticket ID
        in: path
        name: ticketId
        required: true
        type: integer
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
      security:
      - ApiKeyAuth: []
      tags:
      - Ticket
  /api/v1/users:
    get:
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - ApiKeyAuth: []
      tags:
      - User
  /api/v1/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      tags:
      - User
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - ApiKeyAuth: []
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"